fix FORMs instances not being loaded on dotPalette

+++ apps/dotcms-ui/src/app/portlets/dot-edit-page/components/dot-palette/dot-palette-contentlets/dot-palette-contentlets.component.ts
@@ -8,8 +8,9 @@ import {
     ViewChild
 } from '@angular/core';
 import { DotContentletEditorService } from '@dotcms/app/view/components/dot-contentlet-editor/services/dot-contentlet-editor.service';
-import { DotCMSContentlet } from '@dotcms/dotcms-models';
-import { DotPaginatorESContentService, OrderDirection } from '@services/dot-paginator-es-content/dot-paginator-es-content.service';
+import { DotCMSContentlet, DotCMSContentType } from '@dotcms/dotcms-models';
+import { DotPaginatorESContentService } from '@services/dot-paginator-es-content/dot-paginator-es-content.service';
+import { PaginatorService } from '@services/paginator';
 import { LazyLoadEvent } from 'primeng/api';
 import { take } from 'rxjs/operators';
 import { DotPaletteInputFilterComponent } from '../dot-palette-input-filter/dot-palette-input-filter.component';
@@ -25,19 +26,31 @@ export class DotPaletteContentletsComponent implements OnChanges {
 
     @Input() contentTypeVariable: string;
     @Output() hide = new EventEmitter();
-    items: DotCMSContentlet[];
+    items: DotCMSContentlet[] | DotCMSContentType[];
+    isFormContentType: boolean;
     hideNoResults = true;
 
     constructor(
         public paginatorESService: DotPaginatorESContentService,
+        public paginationService: PaginatorService,
         private dotContentletEditorService: DotContentletEditorService
     ) {}
 
     ngOnChanges(changes: SimpleChanges): void {
         if (changes?.contentTypeVariable?.currentValue) {
-            this.paginatorESService.setExtraParams('+contentType', this.contentTypeVariable);
-            this.paginatorESService.deleteExtraParams('+title');
-            this.paginatorESService.paginationPerPage = 15;
+            this.isFormContentType = changes?.contentTypeVariable?.currentValue === 'forms';
+
+            if (this.isFormContentType) {
+                this.paginationService.url = `v1/contenttype`;
+                this.paginationService.paginationPerPage = 15;
+                this.paginationService.sortField = 'modDate';
+                this.paginationService.setExtraParams('type', 'Form');
+                this.paginationService.sortOrder = 1;
+            } else {
+                this.paginatorESService.setExtraParams('+contentType', this.contentTypeVariable);
+                this.paginatorESService.deleteExtraParams('+title');
+                this.paginatorESService.paginationPerPage = 15;
+            }
 
             this.loadData();
         }
@@ -50,14 +63,28 @@ export class DotPaletteContentletsComponent implements OnChanges {
      * @memberof DotAppsConfigurationComponent
      */
     loadData(event?: LazyLoadEvent): void {
-        this.paginatorESService
-            .getWithOffset((event && event.first) || 0)
-            .pipe(take(1))
-            .subscribe((data: any[]) => {
-                this.items = data;
-                this.hideNoResults = !!data?.length;
-            });
-
+        if (this.isFormContentType) {
+            this.paginationService.setExtraParams(
+                'filter',
+                this.filterInput.searchInput.nativeElement.value
+            );
+            this.paginationService
+                .getWithOffset((event && event.first) || 0)
+                .pipe(take(1))
+                .subscribe((data: DotCMSContentType[] | DotCMSContentlet[]) => {
+                    data.forEach((item) => (item.contentType = item.variable = 'FORM'));
+                    this.items = data;
+                    this.hideNoResults = !!data?.length;
+                });
+        } else {
+            this.paginatorESService
+                .getWithOffset((event && event.first) || 0)
+                .pipe(take(1))
+                .subscribe((data: DotCMSContentlet[]) => {
+                    this.items = data;
+                    this.hideNoResults = !!data?.length;
+                });
+        }
     }
 
     paginate(event: LazyLoadEvent) {
@@ -82,11 +109,17 @@ export class DotPaletteContentletsComponent implements OnChanges {
 
     filterContentlets(value: string): void {
         value = value.trim();
-        if (value.indexOf(' ') > 0) {
-            value = `'${value.replace(/'/g, "\\'")}'`;
+
+        if (this.isFormContentType) {
+            this.paginationService.searchParam = 'variable';
+            this.paginationService.filter = value;
+        } else {
+            if (value.indexOf(' ') > 0) {
+                value = `'${value.replace(/'/g, "\\'")}'`;
+            }
+            this.paginatorESService.setExtraParams('+title', `${value}*`);
         }
-        this.paginatorESService.setExtraParams('+title', `${value}*`);
+
         this.loadData();
     }
-
 }
